update local code

+++ src/backend/backend/src/main/java/com/group8/backend/dto/CartItemDTO.java
@@ -1,7 +1,6 @@
 package com.group8.backend.dto;
 
 import lombok.*;
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter @Setter
@@ -12,7 +11,7 @@ public class CartItemDTO {
     private String foodName;
     private Integer quantity;
     private String note;
-    private BigDecimal unitPrice;
-    private BigDecimal totalPrice;
+    private Long unitPrice;
+    private Long totalPrice;
     private List<CartItemToppingDTO> toppings;
 }
