+++ src/backend/backend/src/main/java/com/group8/backend/dto/CartResponseDTO.java
@@ -1,7 +1,6 @@
 package com.group8.backend.dto;
 
 import lombok.*;
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter @Setter
@@ -10,6 +9,6 @@ public class CartResponseDTO {
     private String cartId;
     private String userId;
     private String merchantId;
-    private BigDecimal subtotalPrice;
+    private Long subtotalPrice;
     private List<CartItemDTO> items;
 }
